multi-camera-server/pyproject.toml
[project]
name = "multi_camera_system"
version = "0.0.1"
authors = [
  { name="Tomi Tuhkanen", email="tomi.tuhkanen@iki.fi" },
]
description = "Multi camera system"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "opencv-python",
    "numpy",
    "bleak",
    "fastapi",
    "uvicorn[standard]",
    "psycopg",
    "psycopg-binary",
    "minio",
    "python-dotenv"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "mypy",
    "ruff",
    "pre-commit",
    "httpx"
]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    ".git",
    ".eggs",
    ".venv",
    "__pycache__",
    "build",
    "dist"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "A",     # flake8-builtins
    "C90",   # maccabe
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "G",     # flake8-logging-format
    "I",     # isort
    "PERF",  # perflint
    "PIE",   # flake8-pie
    "PL",    # pylint
    "PTH",   # flake8-use-pathlib
    "Q",     # flake8-quotes
    "RUF",   # ruff-specific rules
    "SIM",   # flake8-simplify
    "W",     # pycodestyle warnings
]
ignore = [
    "PLR2004", # Magic value used in comparison
    "RUF006",  # Store a reference to the return value of asyncio.create_task
    "A001",  # Variable `xxx` is shadowing a Python builtin
    "A002",  # Function argument `xxx` is shadowing a Python builtin
    "PERF203",  # `try`-`except` within a loop incurs performance overhead
    "SIM117",  # Use a single `with` statement with multiple contexts
    "RUF017",  # Avoid quadratic list summation
    "E402",  # Module level import not at top of file
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"  # Allow unused variables when underscore-prefixed.

[tool.ruff.lint.isort]
force-single-line = false
lines-after-imports = 2
lines-between-types = 1
combine-as-imports = true

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = "tests/"
